requirements.txt
crewai>=0.28.0
langchain>=0.0.325
langchain-community>=0.0.10
pandas>=1.5.3
matplotlib>=3.7.1
seaborn>=0.12.2
python-dotenv>=1.0.0
openpyxl>=3.1.0  # For Excel file support
jupyterlab>=4.0.0  # Optional, for development